:sparkles: [blockchain]: Merged juror testing files #121
